+++ scenes/ui/underground_car_meet/ui.gd
@@ -7,6 +7,10 @@ extends UI
 @export var player_cash_label : Label
 @export var negotiation_buttons : Control
 @export var counter_offer_input : SpinBox
+@export var car_preview_sub_viewport : SubViewport
+
+var preview_scene = preload("res://cars/car_preview_scene/car_preview.tscn")
+@onready var preview_scene_instance
 # The accept offer and propose counter-offer button are in a separate array from the decline/cancel
 # button because once a negotiation is cancelled, the accept and counter-offer buttons are meant to
 # be fully disabled whilst the cancel button needs to be enabled to close the menu.
@@ -27,6 +31,7 @@ extends UI
 @export var race_ui : Control
 signal run_finished
 
+var racer_data : Dictionary
 var negotiation_phase : int = 0
 var current_wager : int = 0
 
@@ -37,6 +42,8 @@ signal racer_interaction_menu_closed(can_enable_raycast : bool)
 func _ready():
 	update_labels()
 	race_ui.connect("run_finished", run_finished.emit)
+	preview_scene_instance = preview_scene.instantiate()
+	car_preview_sub_viewport.add_child(preview_scene_instance)
 
 func update_labels():
 	$BaseUI/VBoxContainer/BaseUI.update_labels()
@@ -75,7 +82,7 @@ func _on_propose_counter_offer_button_pressed():
 		_:
 			pass
 
-func show_racer_interact_screen(racer_data : Dictionary):
+func show_racer_interact_screen(_racer_data : Dictionary):
 	## Reset interaction screen to default state
 	racer_interact_screen.show()
 	update_labels()
@@ -93,27 +100,33 @@ func show_racer_interact_screen(racer_data : Dictionary):
 	negotiation_buttons.show()
 	start_race_container.hide()
 	
+	racer_data = _racer_data
+
+	## Display preview of racer car
+	preview_scene_instance.set_racer_car(racer_data["general_car_data"], racer_data["wheels"], racer_data["color"])
+	
 	## Fill with new data
 	var acceleration : float = 0
-	for i in racer_data["car_data"]["acceleration_rate_for_gear"]:
+	for i in racer_data["performance_data"]["acceleration_rate_for_gear"]:
 		acceleration += i
-	acceleration /= racer_data["car_data"]["acceleration_rate_for_gear"].size()
+	acceleration /= racer_data["performance_data"]["acceleration_rate_for_gear"].size()
 	
 	var display_data : String = \
 		"Racer data:
+		Cash: $%d
 		Car: %s %s
-		Top speed: %d
+		Top speed: %d KM/h
 		Max RPM: %d
 		Acceleration: %.2f
 		Has nitrous: %s
 		Rep: %d
 		Wins: %d
-		Losses: %d
-		" % [
-			racer_data["car_data"]["manufacturer"],
-			racer_data["car_data"]["name"],
-			(racer_data["car_data"]["top_speed_mps"]*3.6),
-			(racer_data["car_data"]["max_rpm"]),
+		Losses: %d" % [
+			racer_data["money"],
+			racer_data["general_car_data"]["manufacturer"],
+			racer_data["general_car_data"]["name"],
+			(racer_data["performance_data"]["top_speed_mps"]*3.6),
+			(racer_data["performance_data"]["max_rpm"]),
 			acceleration,
 			"Yes" if racer_data["upgrades"]["nitrous"] > 0 else "No",
 			racer_data["rep"],
@@ -138,24 +151,22 @@ func negotiate_wager(phase : int, offer : int, check_offer : bool):
 					cancel_negotiations()
 					wager_negotiation_labels[6].text = ""
 					current_wager_label.text = "Not enough money to wager for a race ($1000 minimum)"
+				# TODO:
+				#elif PlayerStats.get_fuel() > 0:
+				#	cancel_negotiatons()
+				#	wager_negotiation_labels[6].text = ""
+				#	current_wager_label.text = "Not enough fuel to participate in a race (1 minimum)"
 				else:
-					if PlayerStats.get_cash() > 10000:
-						if PlayerStats.get_cash() > 30000:
-							initial_wager = randi_range(5000, 30000)
-						else:
-							initial_wager = randi_range(5000, PlayerStats.get_cash())
-					elif PlayerStats.get_cash() >= 1000:
-						initial_wager = randi_range(1000, PlayerStats.get_cash())
+					initial_wager = randi_range(int(racer_data["money"] * 0.2), racer_data["money"])
+					
 					wager_negotiation_labels[0].text = "Initial wager: $%d" % initial_wager
 					wager_negotiation_labels[6].text = "Negotiations in progress"
 					current_wager = initial_wager
 					current_wager_label.text = "Current wager: $%d" % initial_wager
 			else: # Consider first player counter-offer
-				# TODO: Use a timer to simulate thinking about the offer
-				# TODO: Add a label that tells the player that the racer is thinking about the offer (change the current_wager_label.text to dots or "Thinking..."
 				if racer_consider_offer(current_wager, offer) == true:
-					racer_accept_offer()
 					current_wager = offer
+					racer_accept_offer()
 				else:
 					var new_offer : int = generate_counter_offer(offer)
 					negotiation_phase += 1
@@ -193,14 +204,26 @@ func negotiate_wager(phase : int, offer : int, check_offer : bool):
 ## True accepts offer, can start racing. False denies offer, makes another offer.
 ## Currently always false to test messages
 func racer_consider_offer(current_wager : int, player_offer : int) -> bool:
-	# TODO: Expand logic for accepting/denying counter-offers
-	# 2 predicates:
-	# Difference between player and racer rep	(Equal, Greater, Lesser)
-	# Difference between player and racer offer	(Equal, Greater, Lesser)
-	# Player is not making any sensible counter-offer, just repeating the offer given by the racer
+	const greed_threshold_percent : int = 10 # Percent above initial wager that the racer may accept
+	const lowball_threshhold_percent : int = 80 # Percent within the initial wager that the racer will accept
+	
 	if current_wager == player_offer:
-		return true
+		return true # Same offer, accept wager
 	else:
+		var greed_threshold : int = int(current_wager * (1 + greed_threshold_percent / 100.0))
+		var min_acceptable_wager : int = int(current_wager * (lowball_threshhold_percent / 100.0))
+		
+		print_debug("Current wager: $", current_wager," | Player offer: $", player_offer, " | Greed threshhold : $", greed_threshold, \
+			" | Min. acceptable wager: $", min_acceptable_wager
+		)
+		
+		if player_offer > current_wager:
+			if player_offer <= greed_threshold && player_offer <= racer_data["money"]:
+				return true
+		else:
+			if player_offer >= min_acceptable_wager :
+				return true
+		
 		return false
 
 func cancel_negotiations():
@@ -213,16 +236,18 @@ func cancel_negotiations():
 func racer_accept_offer():
 	negotiation_buttons.hide()
 	start_race_container.show()
+	current_wager_label.text = "Current wager: $%d" % current_wager
 	wager_negotiation_labels[6].text = "Deal."
 
 func generate_counter_offer(offer : int) -> int:
 	var difference_between_offers : int = abs(offer-current_wager)
-	var smaller_offer : int
-	if current_wager < offer:
-		smaller_offer = current_wager
-	else:
-		smaller_offer = offer
-	return smaller_offer+((difference_between_offers / 2) + (difference_between_offers / 2)*randf_range(.25, .75))
+	var smaller_offer : int = min(current_wager, offer)
+	var counter_offer :int = smaller_offer+((difference_between_offers / 2) + (difference_between_offers / 2)*randf_range(.25, .75))
+	
+	counter_offer = min(counter_offer, racer_data["money"])
+	counter_offer = max(counter_offer, current_wager)
+	
+	return counter_offer
 
 func show_race_ui():
 	base_ui.hide()